+++ libs/vision/src/checkerboard_cam_calib.cpp
@@ -342,9 +342,12 @@ bool mrpt::vision::checkerBoardCameraCalibration(
 		{
 			TImageCalibData& dat = it->second;
 			if (!dat.img_original.isExternallyStored())
-				dat.img_original.undistort(
-					dat.img_rectified, out_camera_params);
-		}  // end undistort
+			{
+				mrpt::img::CImage im;
+				dat.img_original.undistort(im, out_camera_params);
+				dat.img_rectified = std::move(im);
+			}
+		}
 
 		// -----------------------------------------------
 		// Reproject points to measure the fit sqr error